type mistake

+++ main.py
@@ -39,7 +39,7 @@ parser.add_argument('--n_epoch', default=100, type=int)
 parser.add_argument('--beta1', default=0.5, type=float, help='momentum term of adam')
 parser.add_argument('--lambd', default=100.0, type=float, help='weight for L1 loss')
 parser.add_argument('--lambd_d', default=0.5, type=float, help='D loss scale')
-parser.add_argument('--d_update_frequency', default=1.0, type=float, help='discriminator parameter update frequency')
+parser.add_argument('--d_update_frequency', default=1, type=int, help='discriminator parameter update frequency')
 # Files
 parser.add_argument('--out_dir', default='./checkpoints', type=str)
 parser.add_argument('--data_dir', default='./datasets/maps/', type=str)
